Move pseudonym action from Profiles.java to a specific class
