+++ mvc/src/main/java/org/karnak/profilepipe/Profiles.java
@@ -52,15 +52,13 @@ public class Profiles {
     private final Logger LOGGER = LoggerFactory.getLogger(Profiles.class);
 
     private Profile profile;
+    private Pseudonym pseudonymUtil;
     private final ArrayList<ProfileItem> profiles;
     private final Map<String, MaskArea> maskMap;
-    private Cache<String, Patient> cache;
-    private Cache<String, Patient> mainzellisteCache;
 
     public Profiles(Profile profile) {
         this.maskMap = new HashMap<>();
-        this.cache = AppConfig.getInstance().getCache();
-        this.mainzellisteCache = AppConfig.getInstance().getMainzellisteCache();
+        this.pseudonymUtil = new Pseudonym();
         this.profile = profile;
         this.profiles = createProfilesList();
     }
@@ -115,79 +113,6 @@ public class Profiles {
             return profiles;
         }
         return null;
-    }
-
-    private void cachingMainzellistePseudonym(String pseudonym, PatientMetadata patientMetadata) {
-        final Patient patient = new Patient(pseudonym,
-                patientMetadata.getPatientID(),
-                patientMetadata.getPatientFirstName(),
-                patientMetadata.getPatientLastName(),
-                patientMetadata.getLocalDatePatientBirthDate(),
-                patientMetadata.getPatientSex(),
-                patientMetadata.getIssuerOfPatientID());
-        mainzellisteCache.remove(PatientCachingUtil.generateKey(patientMetadata));
-        mainzellisteCache.put(PatientCachingUtil.generateKey(patientMetadata), patient); //new extid
-    }
-
-    public String getMainzellistePseudonym(PatientMetadata patientMetadata, String externalPseudonym, IdTypes idTypes) throws IOException, InterruptedException {
-        final String cachedPseudonym = PatientCachingUtil.getPseudonym(patientMetadata, mainzellisteCache);
-        if (cachedPseudonym != null) {
-            cachingMainzellistePseudonym(cachedPseudonym, patientMetadata);
-            return cachedPseudonym;
-        }
-
-        PseudonymApi pseudonymApi = new PseudonymApi(externalPseudonym);
-        final Fields newPatientFields = patientMetadata.generateMainzellisteFields();
-
-        String pseudonym = pseudonymApi.createPatient(newPatientFields, idTypes);
-        cachingMainzellistePseudonym(pseudonym, patientMetadata);
-        return pseudonym;
-    }
-
-    private String getExtIDInDicom(DicomObject dcm, Destination destination) {
-        if (destination.getIdTypes().equals(IdTypes.ADD_EXTID)) {
-            String cleanTag = destination.getTag().replaceAll("[(),]", "").toUpperCase();
-            final String tagValue = dcm.getString(TagUtils.intFromHexString(cleanTag)).orElse(null);
-            if (tagValue != null && destination.getDelimiter() != null && destination.getPosition() != null) {
-                String delimiterSpec = SpecialCharacter.escapeSpecialRegexChars(destination.getDelimiter());
-                try {
-                    return tagValue.split(delimiterSpec)[destination.getPosition()];
-                } catch (ArrayIndexOutOfBoundsException e) {
-                    LOGGER.error("Can not split the external pseudonym", e);
-                    return null;
-                }
-            } else if (tagValue != null) {
-                return tagValue;
-            }
-        }
-        return null;
-    }
-
-    private String generatePseudonym(Destination destination, DicomObject dcm) {
-        String pseudonym;
-        if (destination.getSavePseudonym() != null && destination.getSavePseudonym() == false) {
-            pseudonym = getExtIDInDicom(dcm, destination);
-            if (pseudonym == null) {
-                throw new IllegalStateException("Cannot get a pseudonym in a DICOM tag");
-            }
-        } else {
-            PatientMetadata patientMetadata = new PatientMetadata(dcm);
-            pseudonym = PatientCachingUtil.getPseudonym(patientMetadata, cache);
-
-            if(pseudonym == null){
-                final String issuerOfPatientID = dcm.getString(Tag.IssuerOfPatientID)
-                        .orElse(profile.getDefaultissueropatientid());
-                patientMetadata.setIssuerOfPatientID(issuerOfPatientID);
-                try {
-                    pseudonym = getMainzellistePseudonym(patientMetadata, getExtIDInDicom(dcm, destination),
-                            destination.getIdTypes());
-                } catch (Exception e) {
-                    LOGGER.error("Cannot get a pseudonym with Mainzelliste API {}", e);
-                    throw new IllegalStateException("Cannot get a pseudonym in cache or with Mainzelliste API");
-                }
-            }
-        }
-        return pseudonym;
     }
 
     public void applyAction(DicomObject dcm, DicomObject dcmCopy, HMAC hmac,
@@ -248,7 +173,7 @@ public class Profiles {
         MDC.put("issuerOfPatientID", IssuerOfPatientID);
         MDC.put("PatientID", PatientID);
 
-        String pseudonym = generatePseudonym(destination, dcm);
+        String pseudonym = pseudonymUtil.generatePseudonym(destination, dcm, profile.getDefaultissueropatientid());
 
         String profilesCodeName = String.join(
                 "-" , profiles.stream().map(profile -> profile.getCodeName()).collect(Collectors.toList())
